pyproject.toml
[project]
name = "vocabuildary"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "apscheduler>=3.11.0",
    "dotenv>=0.9.9",
    "python-telegram-bot>=22.0",
    "schedule>=1.2.2",
]
